+++ plugins/org.jkiss.dbeaver.ext.oceanbase/plugin.xml
@@ -13,9 +13,9 @@
                 dialect="oceanbase">
                 
             <tree
-                  icon="icons/mysql_icon.png"
-                  label="MySQL data source"
-                  path="mysql">
+                  icon="icons/ob_icon.png"
+                  label="OceanBase data source"
+                  path="oceanbase">
                 <folder type="org.jkiss.dbeaver.ext.mysql.model.MySQLCatalog" label="%tree.databases.node.name" icon="#folder_schema" description="%tree.databases.node.tip">
                     <items label="%tree.database.node.name" path="database" property="catalogs" icon="#database">
                         <folder type="org.jkiss.dbeaver.ext.mysql.model.MySQLTable" label="%tree.tables.node.name" icon="#folder_table" description="%tree.tables.node.tip">
@@ -113,8 +113,8 @@
                 <driver
                         id="alipay_oceanbase"
                         label="OceanBase"
-                        class="com.mysql.jdbc.Driver"
-                        sampleURL="jdbc:mysql://{host}:{port}[/{database}]"
+                        class="com.alipay.oceanbase.jdbc.Driver"
+                        sampleURL="jdbc:oceanbase://{host}:{port}[/{database}]"
                         defaultPort="2881"
                         description="Alipay Oceanbase"
                         webURL="https://open.oceanbase.com"
@@ -188,7 +188,9 @@
         </authModel>
     </extension>
     <extension point="org.jkiss.dbeaver.ui.propertyConfigurator">
-        <propertyConfigurator class="org.jkiss.dbeaver.ext.oceanbase.model.auth.OceanbaseAuthModelDatabaseNative" uiClass="org.jkiss.dbeaver.ui.dialogs.connection.DatabaseNativeAuthModelConfigurator"/>
+        <propertyConfigurator 
+        	class="org.jkiss.dbeaver.ext.oceanbase.model.auth.OceanbaseAuthModelDatabaseNative" 
+        	uiClass="org.jkiss.dbeaver.ui.dialogs.connection.DatabaseNativeAuthModelConfigurator"/>
     </extension>
     
 </plugin>
